+++ changelog.md
@@ -1,5 +1,6 @@
 # Changelog
 
 ## 0.1
+* [Added] New parameter values stored into outputdir
 * [Fix] Fixed initial sorting with no initial annotation
 * [Added] Cache images for faster GUI response